requirements.txt
phidata==2.7.10
langchain_openai==0.3.7
duckduckgo_search==7.5.0
yfinance==0.2.54
python-dotenv==1.0.1
slack_bolt==1.22.0
langchain_chroma==0.2.2
pypdf==5.3.1
langchain-text-splitters==0.3.6
langchain_community==0.3.18
langchain==0.3.19
google_generativeai==0.8.4
